rename UserInfoDirectives
